Use float64 dtype for board data.

+++ app/nubrain/experiment/global_config.py
@@ -11,6 +11,6 @@ class GlobalConfig:
         self.stim_start_marker = 1.0
         self.stim_end_marker = 2.0
         # Data type for board data to use when saving to hdf5 file.
-        self.hdf5_dtype = "float32"
+        self.hdf5_dtype = "float64"
         # Resize longest image dimension to this size when saving image to hdf5 file.
         self.max_img_storage_dimension = 128
